+++ src/console_results.sh
@@ -2,7 +2,7 @@
 
 _START_TIME=$(date +%s%N);
 
-function renderResult() {
+function Console::renderResult() {
   local tests_passed=$1
   local tests_failed=$2
   local assertions_passed=$3
@@ -32,27 +32,27 @@ function renderResult() {
   printf " %s total\n" "$total_assertions"
 
   if [[ "$tests_failed" -gt 0 ]]; then
-    printExecutionTime
+    Console::printExecutionTime
     return
   fi
 
   printf "%s%s%s\n" "$_COLOR_ALL_PASSED" "All tests passed" "$_COLOR_DEFAULT"
-  printExecutionTime
+  Console::printExecutionTime
 }
 
-function printExecutionTime() {
+function Console::printExecutionTime() {
   if [ "$_OS" != "OSX" ]; then
     _EXECUTION_TIME=$((($(date +%s%N) - "$_START_TIME") / 1000000))
     printf "${_COLOR_BOLD}%s${_COLOR_DEFAULT}\n" "Time taken: ${_EXECUTION_TIME} ms"
   fi
 }
 
-function printSuccessfulTest() {
+function Console::printSuccessfulTest() {
   local test_name=$1
   printf "%s✓ Passed%s: %s\n" "$_COLOR_PASSED" "$_COLOR_DEFAULT" "${test_name}"
 }
 
-function printFailedTest() {
+function Console::printFailedTest() {
   local test_name=$1
   local expected=$2
   local failure_condition_message=$3