+++ src/helpers.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
 #
-# @param string Eg: "test_some_logic"
+# @param $1 string Eg: "test_some_logic_camelCase"
 #
-# @result string Eg: "Some logic"
+# @result string Eg: "Some logic camelCase"
 #
-function normalizeTestFunctionName() {
+function Helper::normalizeTestFunctionName() {
   local original_function_name="$1"
   local result
 
@@ -21,7 +21,14 @@ function normalizeTestFunctionName() {
   echo "$result"
 }
 
-function getFunctionsToRun() {
+#
+# @param $1 string Eg: "prefix"
+# @param $2 string Eg: "filter"
+# @param $3 array Eg: "[fn1, fn2, prefix_filter_fn3, fn4, ...]"
+#
+# @result array Eg: "[prefix_filter_fn3, ...]" The filtered functions with prefix
+#
+function Helper::getFunctionsToRun() {
   local prefix=$1
   local filter=$2
   local function_names=$3
@@ -50,4 +57,29 @@ function getFunctionsToRun() {
   done
 
   echo "${functions_to_run[@]}"
+}
+
+#
+# @param $1 string Eg: "do_something"
+#
+function Helper::executeFunctionIfExists() {
+  local function_name=$1
+
+  if declare -F | awk '{print $3}' | grep -Eq "^${function_name}$"; then
+    "$function_name"
+  fi
+}
+
+#
+# @param $1 string Eg: "do_something"
+#
+function Helper::unsetIfExists() {
+  local function_name=$1
+
+  if declare -F | awk '{print $3}' | grep -Eq "^${function_name}$"; then
+    unset "$function_name"
+    return 0
+  fi
+
+  return 1
 }